requirements.txt
configparser==5.0.2
PyJWT==2.2.0
Werkzeug==2.0.1
requests==2.26.0
beautifulsoup4==4.10.0
Flask==2.0.1
soupsieve==2.2.1
cryptography==35.0.0
pip==21.2.4
pytz==2021.1
MarkupSafe==2.0.1
Jinja2==3.0.1
click==8.0.1
itsdangerous==2.0.1
certifi==2021.5.30
urllib3==1.26.7
six==1.16.0
idna==3.2
aniso8601==9.0.1
pycparser==2.20
setuptools==41.2.0
psycopg2-binary
flask-restful